+++ sdk/src/Core/Sampling/GetSamplingTargetsResponse.cs
@@ -19,7 +19,7 @@ using System.Collections.Generic;
 namespace Amazon.XRay.Recorder.Core.Sampling
 {
     /// <summary>
-    /// Wrapper for <see cref="AmazonXRayClient.GetSamplingTargetsAsync(Model.GetSamplingTargetsRequest, System.Threading.CancellationToken)"/> API call response.
+    /// Class for keep last rule modification timestamp and list of sampling targets from x-ray backend.
     /// </summary>
     public class GetSamplingTargetsResponse
     {