+++ sdk/src/Handlers/AwsSdk/Internal/AWSXRaySDKUtils.cs
@@ -26,6 +26,13 @@ namespace Amazon.XRay.Recorder.Handlers.AwsSdk.Internal
     {
         private static readonly String XRayServiceName = "XRay";
         private static readonly ISet<String> WhitelistedOperations = new HashSet<String> { "GetSamplingRules", "GetSamplingTargets" };
+
+        // Collection to uniform service names across X-Ray SDKs.
+        private static readonly Dictionary<string, string> FormattedServiceNames = new Dictionary<string, string>()
+        {
+            { "SimpleNotificationService" , "SNS" },
+        };
+
         internal static bool IsBlacklistedOperation(String serviceName, string operation)
         {
             if (string.Equals(serviceName, XRayServiceName) && WhitelistedOperations.Contains(operation))
@@ -33,6 +40,16 @@ namespace Amazon.XRay.Recorder.Handlers.AwsSdk.Internal
                 return true;
             }
             return false;
+        }
+
+        internal static string FormatServiceName(string serviceName)
+        {
+            if (FormattedServiceNames.TryGetValue(serviceName, out string formattedName))
+            {
+                return formattedName;
+            }
+
+            return serviceName;
         }
     }
 }