+++ sdk/src/Handlers/EntityFramework/EFUtil.cs
@@ -1,6 +1,6 @@
 ﻿//-----------------------------------------------------------------------------
 // <copyright file="EFUtil.cs" company="Amazon.com">
-//      Copyright 2016 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//      Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 //      Licensed under the Apache License, Version 2.0 (the "License").
 //      You may not use this file except in compliance with the License.
@@ -15,78 +15,152 @@
 // </copyright>
 //-----------------------------------------------------------------------------
 
-using System.Collections.Generic;
+using System;
 using System.Data.Common;
 using System.Text.RegularExpressions;
-using Microsoft.EntityFrameworkCore;
+using Amazon.Runtime.Internal.Util;
+using Amazon.XRay.Recorder.Core;
+#if NET45
+using Amazon.XRay.Recorder.Core.Internal.Utils;
+#endif
 
 namespace Amazon.XRay.Recorder.Handlers.EntityFramework
 {
     /// <summary>
     /// Utilities for EFInterceptor
     /// </summary>
-    public static class EFUtil
+    internal static class EFUtil
     {
-        private static readonly string DefaultDatabaseType = "EntityFrameworkCore";
+        private static readonly string DefaultDbTypeEntityFramework = "entityframework";
+        private static readonly string SqlServerCompact35 = "sqlservercompact35";
+        private static readonly string SqlServerCompact40 = "sqlservercompact40";
+        private static readonly string MicrosoftSqlClient = "microsoft.data.sqlclient";
+        private static readonly string SystemSqlClient = "system.data.sqlclient";
+        private static readonly string SqlServer = "sqlserver";
 
         private static readonly string[] UserIdFormatOptions = { "user id", "username", "user", "userid" }; // case insensitive
 
-        // Some database providers may not support Entity Framework Core 3.0 and above now
-        // https://docs.microsoft.com/en-us/ef/core/providers/?tabs=dotnet-core-cli
-        private static readonly Dictionary<string, string> DatabaseTypes = new Dictionary<string, string>()
-        {
-            { "Microsoft.EntityFrameworkCore.SqlServer" , "sqlserver" },
-            { "Microsoft.EntityFrameworkCore.Sqlite" , "sqlite" },
-            { "Npgsql.EntityFrameworkCore.PostgreSQL" , "postgresql" },
-            { "Pomelo.EntityFrameworkCore.MySql" , "mysql" },
-            { "FirebirdSql.EntityFrameworkCore.Firebird" , "firebirdsql" },
-
-            { "Microsoft.EntityFrameworkCore.InMemory" , "inmemory" },
-            { "Microsoft.EntityFrameworkCore.Cosmos" , "cosmosdb" },
-            { "Devart.Data.MySql.EFCore" , "mysql" },
-            { "Devart.Data.Oracle.EFCore" , "oracle" },
-            { "Devart.Data.PostgreSql.EFCore" , "postgresql" },
-            { "Devart.Data.SQLite.EFCore" , "sqlite" },
-            { "FileContextCore" , "filecontextcore" },
-            { "EntityFrameworkCore.Jet" , "jet" },
-            { "EntityFrameworkCore.SqlServerCompact35" , "sqlservercompact35" },
-            { "EntityFrameworkCore.SqlServerCompact40" , "sqlservercompact40" },
-            { "Teradata.EntityFrameworkCore" , "teradata" },
-            { "EntityFrameworkCore.FirebirdSql" , "firebirdsql" },
-            { "EntityFrameworkCore.OpenEdge" , "openedge" },
-            { "MySql.Data.EntityFrameworkCore" , "mysql" },
-            { "Oracle.EntityFrameworkCore" , "oracle" },
-            { "IBM.EntityFrameworkCore" , "ibm" },
-            { "IBM.EntityFrameworkCore-lnx" , "ibm" },
-            { "IBM.EntityFrameworkCore-osx" , "ibm" },
-            { "Pomelo.EntityFrameworkCore.MyCat" , "mycat" }
-        };
+        private static readonly string[] DatabaseTypes = { "sqlserver", "sqlite", "postgresql", "mysql", "firebirdsql",
+                                                           "inmemory" , "cosmosdb" , "oracle" , "filecontextcore" ,
+                                                           "jet" , "teradata" , "openedge" , "ibm" , "mycat" , "vfp"};
 
         private static readonly Regex _portNumberRegex = new Regex(@"[,|:]\d+$");
 
+        private static readonly AWSXRayRecorder _recorder = AWSXRayRecorder.Instance;
+
+        private static readonly Logger _logger = Logger.GetLogger(typeof(EFUtil));
+
         /// <summary>
-        /// Extract database_type from <see cref="DbContext"/>.
+        /// Process command to begin subsegment.
         /// </summary>
-        /// <param name="context">Instance of <see cref="DbContext"/>.</param>
-        /// <returns>Type of database.</returns>
-        public static string GetDataBaseType(DbContext context)
+        /// <param name="command">Instance of <see cref="DbCommand"/>.</param>
+        /// <param name="collectSqlQueriesOverride">Nullable to indicate whether to collect sql query text or not.</param>
+        internal static void ProcessBeginCommand(DbCommand command, bool? collectSqlQueriesOverride)
         {
-            string databaseProvider = context?.Database?.ProviderName;
+            _recorder.BeginSubsegment(BuildSubsegmentName(command));
+            _recorder.SetNamespace("remote");
+            CollectSqlInformation(command, collectSqlQueriesOverride);
+        }
 
-            // Need to check if the context and its following parameter is null or not to avoid exception
-            if (string.IsNullOrEmpty(databaseProvider))
+        /// <summary>
+        /// Process to end subsegment
+        /// </summary>
+        internal static void ProcessEndCommand()
+        {
+            _recorder.EndSubsegment();
+        }
+
+        /// <summary>
+        /// Process exception.
+        /// </summary>
+        /// <param name="exception">Instance of <see cref="Exception"/>.</param>
+        internal static void ProcessCommandError(Exception exception)
+        {
+            _recorder.AddException(exception);
+            _recorder.EndSubsegment();
+        }
+
+        /// <summary>
+        /// Builds the name of the subsegment in the format database@datasource
+        /// </summary>
+        /// <param name="command">Instance of <see cref="DbCommand"/>.</param>
+        /// <returns>Returns the formed subsegment name as a string.</returns>
+        private static string BuildSubsegmentName(DbCommand command)
+            => command.Connection.Database + "@" + RemovePortNumberFromDataSource(command.Connection.DataSource);
+
+        /// <summary>
+        /// Records the SQL information on the current subsegment,
+        /// </summary>
+        private static void CollectSqlInformation(DbCommand command, bool? collectSqlQueriesOverride)
+        {
+            // Get database type from DbCommand
+            string databaseType = GetDataBaseType(command);
+            _recorder.AddSqlInformation("database_type", databaseType);
+
+            DbConnectionStringBuilder connectionStringBuilder = new DbConnectionStringBuilder
             {
-                return DefaultDatabaseType;
+                ConnectionString = command.Connection.ConnectionString
+            };
+
+            // Remove sensitive information from connection string
+            connectionStringBuilder.Remove("Password");
+
+            _recorder.AddSqlInformation("connection_string", connectionStringBuilder.ToString());
+
+            // Do a pre-check for UserID since in the case of TrustedConnection, a UserID may not be available.
+            var user_id = GetUserId(connectionStringBuilder);
+            if (user_id != null)
+            {
+                _recorder.AddSqlInformation("user", user_id.ToString());
             }
 
-            string value = null;
-
-            if (DatabaseTypes.TryGetValue(databaseProvider, out value))
+            if (ShouldCollectSqlText(collectSqlQueriesOverride))
             {
-                return value;
+                _recorder.AddSqlInformation("sanitized_query", command.CommandText);
             }
 
-            return databaseProvider;
+            _recorder.AddSqlInformation("database_version", command.Connection.ServerVersion);
+        }
+
+        /// <summary>
+        /// Extract database_type from <see cref="DbCommand"/>.
+        /// </summary>
+        /// <param name="command">Instance of <see cref="DbCommand"/>.</param>
+        /// <returns>Type of database.</returns>
+        internal static string GetDataBaseType(DbCommand command)
+        {
+            var typeString = command?.Connection?.GetType()?.FullName?.ToLower();
+
+            if (string.IsNullOrEmpty(typeString))
+            {
+                _logger.DebugFormat("Can't extract database type from connection, setting it as default: ({0})", DefaultDbTypeEntityFramework);
+                return DefaultDbTypeEntityFramework;
+            }
+
+            if (typeString.Contains(MicrosoftSqlClient) || typeString.Contains(SystemSqlClient))
+            {
+                return SqlServer;
+            }
+
+            if (typeString.Contains(SqlServerCompact35))
+            {
+                return SqlServerCompact35;
+            }
+
+            if (typeString.Contains(SqlServerCompact40))
+            {
+                return SqlServerCompact40;
+            }
+
+            foreach (var databaseType in DatabaseTypes)
+            {
+                if (typeString.Contains(databaseType))
+                {
+                    return databaseType;
+                }
+            }
+
+            return typeString;
         }
 
         /// <summary>
@@ -94,17 +168,17 @@ namespace Amazon.XRay.Recorder.Handlers.EntityFramework
         /// </summary>
         /// <param name="builder">Instance of <see cref="DbConnectionStringBuilder"/>.</param>
         /// <returns></returns>
-        public static object GetUserId(DbConnectionStringBuilder builder)
+        internal static string GetUserId(DbConnectionStringBuilder builder)
         {
-            object value = null;
             foreach (string key in UserIdFormatOptions)
             {
-                if (builder.TryGetValue(key, out value))
+                if (builder.TryGetValue(key, out object value))
                 {
-                    break;
+                    return (string)value;
                 }
             }
-            return value;
+
+            return null;
         }
 
         /// <summary>
@@ -112,9 +186,17 @@ namespace Amazon.XRay.Recorder.Handlers.EntityFramework
         /// </summary>
         /// <param name="dataSource">The data source.</param>
         /// <returns>The data source string with port number removed.</returns>
-        public static string RemovePortNumberFromDataSource(string dataSource)
+        private static string RemovePortNumberFromDataSource(string dataSource)
         {
             return _portNumberRegex.Replace(dataSource, string.Empty);
         }
+
+#if !NET45
+        private static bool ShouldCollectSqlText(bool? collectSqlQueriesOverride)
+            => collectSqlQueriesOverride ?? _recorder.XRayOptions.CollectSqlQueries;
+#else
+        private static bool ShouldCollectSqlText(bool? collectSqlQueriesOverride)
+            => collectSqlQueriesOverride ?? AppSettings.CollectSqlQueries;
+#endif
     }
 }